// src/features/records/types.ts
export interface PersonalRecord {
  id?: number;

  exerciseId: string;
  exerciseName: string;

  weight: number;
  reps: number;

  estimated1RM: number;

  achievedAt: string;
}
export interface PersonalRecord {
  id?: number;

  exerciseId: string;
  exerciseName: string;

  weight: number;
  reps: number;

  estimated1RM: number;
  achievedAt: string;
}